#570: Extract the Driver Overview fetch & transform logic to clarify responsibilities
