+++ src/app/parcels/ActionBar/ActionButtons/DriverOverview/getDriverOverviewData.ts
@@ -1,26 +1,35 @@
-"use client";
-
-import React from "react";
-import supabase from "@/supabaseClient";
 import { Schema } from "@/databaseUtils";
-import PdfButton from "@/components/FileGenerationButtons/PdfButton";
-import DriverOverviewPdf, {
-    DriverOverviewTablesData,
-    DriverOverviewRowData,
-    DriverOverviewCollectionCentreData,
-} from "@/pdf/DriverOverview/DriverOverviewPdf";
+import supabase from "@/supabaseClient";
 import { logErrorReturnLogId } from "@/logger/logger";
 import { displayNameForDeletedClient, formatDateStringAsDate } from "@/common/format";
-import { Dayjs } from "dayjs";
-import { ParcelsTableRow } from "@/app/parcels/parcelsTable/types";
-import { FileGenerationDataFetchResponse } from "@/components/FileGenerationButtons/common";
 
-interface DriverOverviewData {
-    driverName: string | null;
-    date: Date;
-    tableData: DriverOverviewTablesData;
-    message: string;
+export interface DriverOverviewRowData {
+    name: string;
+    address: {
+        line1: string;
+        line2: string | null;
+        town: string | null;
+        county: string | null;
+        postcode: string | null;
+    };
+    contact?: string;
+    packingDate: string | null;
+    instructions?: string;
+    clientIsActive: boolean;
+    numberOfLabels: number | null;
+    collectionCentre: string;
+    isDelivery: boolean;
 }
+
+export type DriverOverviewTablesData = {
+    collections: DriverOverviewCollectionCentreData[];
+    deliveries: DriverOverviewRowData[];
+};
+
+export type DriverOverviewCollectionCentreData = {
+    collectionCentreName: string;
+    rowData: DriverOverviewRowData[];
+};
 
 type ParcelForDelivery = Schema["parcels"] & {
     client: Schema["clients"];
@@ -39,6 +48,21 @@ type ParcelsForDeliveryResponse =
       };
 
 type ParcelsForDeliveryErrorType = "parcelFetchFailed" | "noMatchingClient" | "noCollectionCentre";
+
+type DriverPdfResponse =
+    | {
+          data: DriverOverviewTablesData;
+          error: null;
+      }
+    | {
+          data: null;
+          error: { type: DriverPdfErrorType; logId: string };
+      };
+
+type DriverPdfErrorType = ParcelsForDeliveryErrorType;
+
+export type DriverOverviewErrorType = DriverPdfErrorType | "driverMessageFetchFailed";
+export type DriverOverviewError = { type: DriverOverviewErrorType; logId: string };
 
 const getParcelsForDelivery = async (parcelIds: string[]): Promise<ParcelsForDeliveryResponse> => {
     const { data, error } = await supabase
@@ -88,18 +112,6 @@ const getParcelsForDelivery = async (parcelIds: string[]): Promise<ParcelsForDel
 
     return { data: dataWithNonNullClients, error: null };
 };
-
-type DriverPdfResponse =
-    | {
-          data: DriverOverviewTablesData;
-          error: null;
-      }
-    | {
-          data: null;
-          error: { type: DriverPdfErrorType; logId: string };
-      };
-
-type DriverPdfErrorType = ParcelsForDeliveryErrorType;
 
 const transformRowToDriverOverviewTableData = (
     parcel: ParcelForDelivery
@@ -162,70 +174,4 @@ const getDriverPdfData = async (parcelIds: string[]): Promise<DriverPdfResponse>
     return { data: tableData, error: null };
 };
 
-interface Props {
-    parcels: ParcelsTableRow[];
-    driverName: string | null;
-    date: Dayjs;
-    onPdfCreationCompleted: () => void;
-    onPdfCreationFailed: (error: DriverOverviewError) => void;
-    disabled: boolean;
-}
-
-export type DriverOverviewErrorType = DriverPdfErrorType | "driverMessageFetchFailed";
-export type DriverOverviewError = { type: DriverOverviewErrorType; logId: string };
-
-const DriverOverviewPdfButton = ({
-    parcels,
-    driverName,
-    date,
-    onPdfCreationCompleted,
-    onPdfCreationFailed,
-    disabled,
-}: Props): React.ReactElement => {
-    const fetchDataAndFileName = async (): Promise<
-        FileGenerationDataFetchResponse<DriverOverviewData, DriverOverviewErrorType>
-    > => {
-        const parcelIds = parcels.map((parcel) => {
-            return parcel.parcelId;
-        });
-        const { data: driverPdfData, error: driverPdfError } = await getDriverPdfData(parcelIds);
-        if (driverPdfError) {
-            return { data: null, error: driverPdfError };
-        }
-        const { data: driverMessageData, error: driverMessageError } = await supabase
-            .from("website_data")
-            .select("name, value")
-            .eq("name", "driver_overview_message")
-            .single();
-        if (driverMessageError) {
-            const logId = await logErrorReturnLogId("Error with fetch: Driver overview message", {
-                error: driverMessageError,
-            });
-            return { data: null, error: { type: "driverMessageFetchFailed", logId: logId } };
-        }
-        return {
-            data: {
-                fileData: {
-                    driverName: driverName,
-                    date: date.toDate(),
-                    tableData: driverPdfData,
-                    message: driverMessageData.value,
-                },
-                fileName: "DriverOverview.pdf",
-            },
-            error: null,
-        };
-    };
-    return (
-        <PdfButton
-            fetchDataAndFileName={fetchDataAndFileName}
-            pdfComponent={DriverOverviewPdf}
-            onFileCreationCompleted={onPdfCreationCompleted}
-            onFileCreationFailed={onPdfCreationFailed}
-            disabled={disabled}
-            formSubmitButton={true}
-        />
-    );
-};
-
-export default DriverOverviewPdfButton;
+export default getDriverPdfData;
